开发基本完成

+++ market-api-parent/market-api-core/src/main/java/com/renqi/market/entity/CustomerTask.java
@@ -39,7 +39,7 @@ public class CustomerTask {
 
     private String taskType;
 
-    private Integer goodId;
+    private Long goodId;
 
     private String goodLinkUrl;
